requirements.txt
3==0.0.0
astroid==2.4.2
autopep8==1.5.4
boto==2.49.0
boto3==1.14.55
botocore==1.17.55
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
docutils==0.15.2
idna==2.10
isort==5.4.2
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql-connector-python==8.0.21
protobuf==3.13.0
pycodestyle==2.6.0
pylint==2.6.0
python-dateutil==2.8.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
toml==0.10.1
urllib3==1.25.10
wrapt==1.12.1
